Changed randomization to use daytime, instead of custom counter.
This uses the new modulo function to remove the need for a counter 
incrementing every tick.

This also removes the need for the #kick_message and #sleep_message 
players, both being replaced by the already existing #query_time.

This also changes how #sleep_messages ops.sleep and #kick_messages 
ops.sleep so that they should be the number of sleep/kick messages, not 
n-1.